requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coverage==7.2.7
cryptography==41.0.2
defusedxml==0.7.1
dj-static==0.0.6
Django==4.2.3
django-stdimage==6.0.1
gunicorn==21.2.0
idna==3.4
model-mommy==2.0.0
mysql==0.0.3
mysql-client==0.0.1
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.7.1
urllib3==2.0.4
